Updated Evidence View Template
We were using the tags inline layout for evidence tags, but as it turns out, the actual tags have much more many characters than anticipated, therefore we had to update this and use an adaptive styling. Also we had to remove the filename column as it doesn't seem to provide that much value to the user.