Fixed client-side  name attribute suffix issue
Because form ecription type `multipart/form-data` required for document upload doesn't take any input attributes other than name,
to be able to submit multiple documents, we need a unique `name` for each input. Initially we reused the same functionality as for creating multiple inputs, but this was generating unique `id`'s for each input.
With this fix we added some logic that checks for the form type and changes the attribute that needs the unique suffix, while not breaking the functionality of other components sharing this script.