Add middleware to post details to API
Middleware checks partial validation and then
checks if there is a send function. If it is then
it is assumed this is the final step of the
journey.

If the send function is successful then the user
is redirected to the next path.

If the send fails then the errors are presented to
the user.