Swap around status and stage filter boxes
In order to view the stage filter box (left-hand column) within
Investment Projects (initial view) the user has to scroll towards the
bottom of the page, Dave Matthews mentioned he would like to see it at
the top of the page due to its importance. The status filter box holds
less importance so this can be swapped with the stage filter box.

In short, both filter boxes are trading positions based on importance.