Made Contact Name a Link On Investment Details Pg
In the 'Investment Projects' section made the client contact name a
link so it could allow navigation to the detailed information of a
contact. See IPBETA-118

Updated tests in response to client_contact change

Because the client_contacts property has to be an object now (so it can
be rendered as a link) the unit tests reqd updating

fixup! fixup! fixup! Updated tests in response to client_contact change