Removing the document nav item when a project code starts with 'DHP'
Currently there are 3 types of project codes, the oldest is 1 and
the latest is 3:

1. 016020 (really old)
2. P-30016857 (CDMS)
3. DHP-00000007 (Data Hub)

At present the document nav item is displayed for all 3 scenarios. This
work excludes the document nav item when the project is new, all new
projects have a code that start with 'DHP'. The documents nav item is
specifically and solely for the S3 migrated documents that are
associated with project codes such as 1 & 2.