DST-1259 Integrate Introspect endpoint with permitted applications
On top of regular user permissions there is a need to have permitted
applications vetting, such that not all users should see all
applications in the global header nav bar, but just the ones they have
been given access to.

For acceptance tests we're using a mock service that doesn't contain the
`Find exporters` key, so we can check if it behaves as expected. as a
fallback, the nav will render all links