Adding three new sub-routes into investment projects.
Renaming /investment-projects to /investments so the app routes now
look like this:

 * /investments/projects (contains the projects list)
 * /investments/profiles (new functionality - coming next)
 * /investments/opportunities (new functionality - coming much later)

Each of these pages will sit in a tabbed component
[ Projects | Profiles | Opportunities]

I've temporarily removed the tabs within /investments/projects
until the /investments/profiles implementation is complete, this is next
with /investments/opportunities coming much later. Essentially we will
unveil (so to speak) a new tab (in order left-to-right) as the features
are implemented.

The skeleton views and controllers are both defined for
/investments/profiles and /investments/opportunities where each view has
a tabbed navigation - not to be confused with the new horizontal local
nav which is currently being rolled out.

Any request for /investments will be redirected to /investments/projects
which is the the default view within investments.

https://uktrade.atlassian.net/browse/IPBETA-238
https://uktrade.atlassian.net/browse/IPBETA-239