Adding three new sub-routes into investment projects.
Renamed /investment-projects to /investments so the app routes look like:

 * /investments/projects (contains the projects list)
 * /investments/profiles (new functionality - coming next)
 * /investments/opportunities (new functionality - coming much later)

Each of these pages will sit in a tabbed component
[ Projects | Profiles | Opportunities]

I've temporarily removed the tabs within /investments/projects
until the /investments/profiles implementation is complete, this is next
with /investments/opportunities coming much later.

Added paths to res.local.paths so we have a single source of truth in
our middleware and within our views making it also easier to maintain
and update.