Adding local navigation app selection for nested apps
A nested app is defined as: /companies/id/investments

The investments nested app has additional nested apps such as
projects, large-capital-profile and growth-capital-profile.

1. /companies/id/investments/projects
2. /companies/id/investments/large-capital-profile
3. /companies/id/investments/growth-capital-profile

Currently the local navigation is only selected on 1, where 2 & 3 are
not selected when navigating to them.

This fix address both issues by introducing a subPath array on a local
nav item.