Adding a "Growth Capital Profile" tab into Companies/Investment
When a user selects the "Growth capital profile" tab we will make an
API call to a new endpoint although this is further down the line as
it's not a top priority.

For now we can just show the "Create a profile" button without any
API call or logic being applied.

https://uktrade.atlassian.net/browse/IPBETA-373