Simplify regex for Proposition Abandoning details textarea
A business need requires users to sometimes add a url with evidence to
why a proposition is being abandoned. Users therefore expect to see an
anchor tag in the summary view, and in order to achieve this we used a
regex, however this regex has unwanted side-effects and we're
simplifying it.

Extra tests added to make sure it still behaves as expected.