Add an `npm` command to run selected unit tests
I would like to be able to run selected unit tests rather than the whole
suite, so that we can do TDD more effectively.

I was going to edit the existing `test:unit` command but as this is used
by a couple of other commands where variables are passed in, it seems
like we need a new command.

I'm not a master at shell commands so mainly copied this from SO[1] -
maybe there's a nicer way of doing it?

[1]https://stackoverflow.com/questions/51401352/if-else-on-arguments-in-npm-run-script