remove unused `eventViewRecord`
We are about to start working on surfacing new data around
Events/Interactions/Service Deliveries, so we want to tidy up the code
before starting to make changes.

As far as I can see we do not use the `eventViewRecord` anywhere. The `getEventDetails` middleware is only needed in the attendees router, when we get the event in `renderAttendees`.