DET-85 Opensearch query for activities by Contact
We want to show all activities from Activity Stream for a Contact under
the Contact Interactions tab, so that we can surface Aventri data for
users.

This commit adds a function to build the opensearch query we'll use to
return activites based on contact. The data hub types don't have the
`object.dit:emailAddress` field on them, so we're having to split them
out and search by contactId.

Co-Authored-By: Onyela Ogah <83657534+onyela-o@users.noreply.github.com>
Co-Authored-By: Edmond DeLos Reyes <28296624+dredmonds@users.noreply.github.com>
Co-Authored-By: Harriet H-W <16647486+Harriethw@users.noreply.github.com>
Co-authored-by: Chloe Wong <chloehkwong1@users.noreply.github.com>
Co-Authored-By: Christopher Sunkel <36161814+cgsunkel@users.noreply.github.com>