DET-85 Add a controller function for activities by contact
We want to show all activities from Activity Stream for a Contact under
the Contact Interactions tab, so that we can surface Aventri data for
users.

This commit adds a function that will be called by a get request from
the Contact Interactions tab, and sends the opensearch query to the data
hub Activity Feed endpoint, which will then send it to Activity Stream.

Co-Authored-By: Onyela Ogah <83657534+onyela-o@users.noreply.github.com>
Co-Authored-By: Edmond DeLos Reyes <28296624+dredmonds@users.noreply.github.com>
Co-Authored-By: Harriet H-W <16647486+Harriethw@users.noreply.github.com>
Co-authored-by: Chloe Wong <chloehkwong1@users.noreply.github.com>
Co-Authored-By: Christopher Sunkel <36161814+cgsunkel@users.noreply.github.com>