DET-85 Add new route to get activities data for contact
We want to show all activities from Activity Stream for a Contact under
the Contact Interactions tab, so that we can surface Aventri data for
users.

This commit adds a new `contacts/{contact-id}/activity/data` end point.
This copies the current pipeline of retrieving data for companies from
Activity Stream (`companies/{company-id}/activity/data`).

We will shortly use this endpoint to send opensearch queries to the
backend, which will then call Activity Stream with that query.

Co-Authored-By: Onyela Ogah <83657534+onyela-o@users.noreply.github.com>
Co-Authored-By: Edmond DeLos Reyes <28296624+dredmonds@users.noreply.github.com>
Co-Authored-By: Harriet H-W <16647486+Harriethw@users.noreply.github.com>
Co-authored-by: Chloe Wong <chloehkwong1@users.noreply.github.com>
Co-Authored-By: Christopher Sunkel <36161814+cgsunkel@users.noreply.github.com>