DET-85 Add `AventriAttendee` React component
We want to show all activities from Activity Stream for a Contact under
the Contact Interactions tab, so that we can surface Aventri data for
users.

This commit adds a new type of component to the activities to render
Aventri Attendee object types from Activity Stream.

This will be behind a feature flag, as we don't currently have final
designs for the component and it's a work in progress.

 Co-Authored-By: Onyela Ogah
<83657534+onyela-o@users.noreply.github.com> Co-Authored-By: Edmond
DeLos Reyes <28296624+dredmonds@users.noreply.github.com>
Co-Authored-By: Harriet H-W
<16647486+Harriethw@users.noreply.github.com> Co-authored-by: Chloe Wong
<chloehkwong1@users.noreply.github.com> Co-Authored-By: Christopher
Sunkel <36161814+cgsunkel@users.noreply.github.com>