DET-99 prepare activity feed sandbox environment
We want users to be able to paginate through the activities that are shown for a Contact (following on from work in DET-85).

This commit returns different fixtures depending on whether the contact query is asking for the second page, so that we can test that different data is shown when using pagination.

Co-Authored-By: Edmond DeLos Reyes <28296624+dredmonds@users.noreply.github.com>
Co-Authored-By: Chloe Wong <70902973+chloehkwong1@users.noreply.github.com>
Co-Authored-By: Onyela Ogah <83657534+onyela-o@users.noreply.github.com>
Co-Authored-By: Christopher Sunkel <36161814+cgsunkel@users.noreply.github.com>