DET-99 Add pagination to ContactInteractionsApp
We want users to be able to paginate through the activities that are shown for a Contact (following on from work in DET-85).

This commit adds redux structure to paginate on `ContactInteractionsApp`. Including the Collection Header with total activities and a `page x of x` row.

The design is still WIP. We tried to re-use existing components like the `CollectionHeader` - it's very close to the `CollectionList` component but we need to iterate through `activities` rather than generic items, and as the designs are changing we thought better to duplicate for now.

Co-Authored-By: Edmond DeLos Reyes <28296624+dredmonds@users.noreply.github.com>
Co-Authored-By: Chloe Wong <70902973+chloehkwong1@users.noreply.github.com>
Co-Authored-By: Onyela Ogah <83657534+onyela-o@users.noreply.github.com>
Co-Authored-By: Christopher Sunkel <36161814+cgsunkel@users.noreply.github.com>