Use default codeql workflow
The current recommended default action for codeql is a bit simpler than
the one we have at the moment.
https://github.com/github/codeql-action#usage

I've removed the cron scheduling because it didn't seem necessary to run
it at 5AM every Tuesday if we are looking at it on PRs/push to master.

We also no longer want to query for `security-and-quality` as we found
it was surfacing too many false positives.