DET-123 Add constants for Contact Activity page sort by
We want to sort Contact Activities by 'newest' or 'oldest' - based on
when the activity was published. This might not be the final sorting
methods we use, but as we only have limited data on activities for
Aventri this makes sense for now.

This commit adds constants for the sort options we'll use coming up. At
some point we will want to do a refactor so that Acitivty Feed stuff is
not under the 'companies' folder, as it's being used across contacts and
in the future probably Events, but that could be in a future PR.

Co-Authored-By: Onyela Ogah <83657534+onyela-o@users.noreply.github.com>