DET-123 Add sort order to Activity Feed controller for Contacts
We want to sort Contact Activities by 'newest' or 'oldest' - based on
when the activity was published. This might not be the final sorting
methods we use, but as we only have limited data on activities for
Aventri this makes sense for now.

This commit adds a sort order for sorting on published to the controller function and DSL query

Co-Authored-By: Onyela Ogah <83657534+onyela-o@users.noreply.github.com>