DET-123 setup Activity Feed sandbox to look for sort order
We want to sort Contact Activities by 'newest' or 'oldest' - based on
when the activity was published. This might not be the final sorting
methods we use, but as we only have limited data on activities for
Aventri this makes sense for now.

This commit configures the Activity Feed sandbox to look at the sort options and return data based on it, so that we can then assert on that data in tests.

Co-Authored-By: Onyela Ogah <83657534+onyela-o@users.noreply.github.com>