DET-123 state management for sort by and page number
We want to sort Contact Activities by 'newest' or 'oldest' - based on
when the activity was published. This might not be the final sorting
methods we use, but as we only have limited data on activities for
Aventri this makes sense for now.

This commit re-configures the state management to use the url params for page number and sortby, and pass them to the api call in the task.

Because the sort component includes a page count based on the route, we no longer need to handle it in redux, so that action can be deleted.

Co-Authored-By: Onyela Ogah <83657534+onyela-o@users.noreply.github.com>