DET-123: Add Sort By component
We want to sort Contact Activities by 'newest' or 'oldest' - based on
when the activity was published. This might not be the final sorting
methods we use, but as we only have limited data on activities for
Aventri this makes sense for now.

This commit uses the `CollectionSort` component to handle the sort by and page count, along with a functional test.

Co-Authored-By: Onyela Ogah <83657534+onyela-o@users.noreply.github.com>