Add `userFeatures` to global React store
We have a story coming up where we will need to check a user's feature
flags on the Events collection page. Because the current function for
checking user feature flags is done via the Express routing middleware,
we need a way to check the flags when routing is done via React.

This commit adds a key to the global store so that React components will
be able to map it to their props. e.g. a mapState2Props function would
look in the state and pass the value to component props like so:

```
...

 const isContactActivitiesFeatureOn =
    state.userFeatureFlags['user-contact-activities']

return {
 ...state[ID],
 isContactActivitiesFeauteOn,
}