use `configureStore` in CompanyLocalHeader
The previously-used `createStore`[1] has been deprecated.

This commit updates the code to use the recommended `configureStore`
function which has an easier to use interface, and installs the
recommended reduxjs/toolkit npm library. It also means we don't need to
use a couple of `redux` core functions `applyMiddleware` and
`combineReducers` as the new toolkit function does those for us.

more info about the switch to redux toolkit here:
https://github.com/reduxjs/redux/blob/93d188673fbcc18d19d645d6ae785a18f534e1d6/index.d.ts#L377
and here:
https://redux.js.org/introduction/why-rtk-is-redux-today#what-does-redux-toolkit-do

[1] https://redux.js.org/api/createstore
[2] https://redux-toolkit.js.org/api/configureStore