use `legacy_createStore` in cypress provider
The previously-used `createStore`[1] has been deprecated.

Elsewhere in the app we are able to use the recommended `configureStore` [2], but in the cypress specs there's some conditional passing of state to the reducer which I'm not sure how to replicate with the new function.

more info about the switch to redux toolkit here:
https://github.com/reduxjs/redux/blob/93d188673fbcc18d19d645d6ae785a18f534e1d6/index.d.ts#L377
and here:
https://redux.js.org/introduction/why-rtk-is-redux-today#what-does-redux-toolkit-do

[1] https://redux.js.org/api/createstore
[2] https://redux-toolkit.js.org/api/configureStore