wip set user features in locals
This decouples the setting of `userfeatures` in `res.locals` from
checking whether a particular flag is enabled. Like with the
environment-level flags, we are now setting the user feature flags in
`res.locals` when the express app is created.

This will allow us (in the next commit) to pass those user feature flags
to the global store for React to use, without having to go through
express router middleware.