Add `CheckUserFeatureFlag` React component
We want to be able to check user feature flags in React land without
going through Express middleware. e.g. for Events Collection

This adds a component that could be used to abstract the logic to get a
user's active features, and return a boolean. So that, for example, we
could decide whether to show a new component.