DET-191: show feature flag in route
We want to show the feature flag under test in the url, so that we know
that it's on.

Note: when this runs first, it throws a warning:
```
Warning: Cannot update
a component (`CheckUserFeatureFlag`) while rendering a different
component (`Task`). To locate the bad setState() call inside `Task`
```
It seems that this is occuring because when we change the history
react-router-dom is updating state for the `CheckUserFeatureFlag`
component (and, I think, its children?) while the child `Task` component
is still rendering, or possibly also trying to update state.

I've followed the advice here [1] so that we don't change the history
until the `Task` is theorectically done, but the error is still
happening, and I think it might be a result of redux. As it isn't
causing any errors atm and this seems too big to solve without
untangling the architecture of the Task component, I'm going to leave it

[1] https://github.com/facebook/react/issues/18178#issuecomment-602323184