TET-227: wait for typeahead to contain result
We only want to start selecting a typeahead result when it has finished
loading the options. The `contains` function in crypress will retry if
it doesn't find it in the DOM, and setting a long timeout will hopefully
mean it waits for a bit, allowing the typeahead to catch up from any
throttling. try again if