TET-227: mock network request to fix flakey test
The `events/fitler-spec` test was flakey because the command to select the first typeahead option was being triggered before the option had rendered, due to the data being throttled/network latency.

This is resolved by intercepting the network request and directly returning mock data, rather than making a 'real' request to the sandbox. We aren't losing any test value here because it's still mocking the same logic as before, just without network latency.