TET-223: Refactor `investments/profiles-filtered-spec` spec
There seemed to be a lot of repetition in this test that didn't seem to
be covering anything new - e.g. once we have tested that a country name
is sent to the backend and shown in chips etc., we know it's succesfully
covered the Task logic, which just sends the query to the API. That
logic won't change if we test with another country, as far as I can
tell. The code coverage report says that coverage hasn't changed.