CPS-384: handle error when no Lead ITA on global HQ
Users were seeing the page hang when trying to add a Lead ITA to a company that was a subsidiary of a global headquarters. It should not be possible to do that, as only Lead ITAs on a global HQ can be added to the subsidiary company.
This is a quick fix to handle the error more gracefully.