Issue ID: DAC_Focus_Order_01
When a keyboard user navigating the service using the tab key activates the ‘Companies’ link in the header of the landing page, they are then taken to the ‘Companies page’. However, when the user clicks ‘Tab’, they are taken to the pagination links at the bottom of the page. This is due to the link for ‘Companies’ being the same as the href="#" of the 1st page pagination link. Removed the link for the page the user is on to stop this behaviour.