Issue ID: DAC_Fieldset_And_Legend_01 refactor
Initial implementation didn't get applied to adding a company. Refactored to be a ternary.

Co-Authored-By: Onyela Ogah <83657534+onyela-o@users.noreply.github.com>
Co-Authored-By: Barbora Kaminskas <106600778+babskami@users.noreply.github.com>

Test spec amended now flieldset and legend added to forms with labels

Co-Authored-By: Onyela Ogah <83657534+onyela-o@users.noreply.github.com>
Co-Authored-By: Barbora Kaminskas <106600778+babskami@users.noreply.github.com>

Radio button assertion amended now flieldset and legend added to forms with labels

Co-Authored-By: Onyela Ogah <83657534+onyela-o@users.noreply.github.com>
Co-Authored-By: Barbora Kaminskas <106600778+babskami@users.noreply.github.com>

Test spec amended now flieldset and legend added to forms with labels

Co-Authored-By: Onyela Ogah <83657534+onyela-o@users.noreply.github.com>
Co-Authored-By: Barbora Kaminskas <106600778+babskami@users.noreply.github.com>

Test spec amended now flieldset and legend added to forms with labels

Co-Authored-By: Onyela Ogah <83657534+onyela-o@users.noreply.github.com>
Co-Authored-By: Barbora Kaminskas <106600778+babskami@users.noreply.github.com>

Further test changes to reflect that fieldset and legends are widely used in inputs

Mostly these changed reflect the need to move up to the parent level before continuing with the test as the label is now contained within a legend tag

Co-Authored-By: Onyela Ogah <83657534+onyela-o@users.noreply.github.com>
Co-Authored-By: Barbora Kaminskas <106600778+babskami@users.noreply.github.com>

Amendments to testing to accommodate fieldset and legend

Co-Authored-By: Onyela Ogah <83657534+onyela-o@users.noreply.github.com>
Co-Authored-By: Barbora Kaminskas <106600778+babskami@users.noreply.github.com>