Fixes a known issue when creating a new interaction
If a user selects a theme (Export, Investment, Trade agreement, Other)
then continues to the next step and partially completes the form
before changing their mind and returns to change the theme, at this point,
we need to clean up after ourselves before the user goes back to the form.

If we do not clean up, then the form values object contains keys and values
relevant to the original form interaction, as the user completes the form
some keys and values are overwritten, others are not, the latter causes
problems with API validation when saving the form as the wrong keys are
sent as part of the payload, which ultimately means the user cannot save
the form (HTTP 400), meaning they have to start over - very frustrating.
Therefore, the cleanest approach is to reset the fields within the form
the moment the users lands on the "Add interaction ..." page: