Improve the NewWindowLink component so it is aligned with GDS and is accessible
Currently, when a screen-reader comes across a link which opens in a new tab the
aria-label="Opens in a new window or tab" overrides the visual label of the link.

That means screen-reader users don’t know what the link refers to and only hears
"Opens in a new window or tab" which is unhelpful.

There appears to be a misunderstanding in how this attribute works on a link amongst
devs including myself, by improving the docs and some minor refactoring it should be
a lot clearer going forward.