When the form unmounts save the previous state
When filling out a form and then redirecting to add a contact we want to
persist the previous form's state so when we save the contact and return
to the form its state is maintained.

This new approach means we no longer have to write to window.sessionStorage
to save the form state and then read from it after saving the contact.

All of this is made possible by using react-router, a soft redirect compared
to the current hard redirect which sets window.location.href.