Fixes an export win pagination bug
When posting to an endpoint using Axios any value that is null in
the payload is stripped out. This is problematic for export wins
paginated pages.

1. When we want to fetch rejected wins we use the payload { confirmed: false}
2. When we want to fetch won wins we use the payload { confirmed: true}
3. When we want to fetch sent wins we use the payload { confirmed: null}

The last API call ultimately fails.

The temporary fix is to cast null to a string "null". The proper fix is
to create an axios intereceptor which we don't have time for - tech-debt!