Specifically target the request to test.
For context, when visiting the index of a page which has default query parameters (and the index url does not contain these), the app makes a second request with those parameters. As a result, multiple requests were caught by the intercept.

This commit changes the intercept statements for most tests to target a specific request which assertions can then be carried out on.