Export project interactions `Sort by` functionality (#7500)
* Layout test scenarios - Sort export project interaction list

Sort a default company export interaction list

* Added test coverage for company export interaction

* Added export interaction collection list test

* Update sandbox server `/v4/export/:exportId` endpoint error response

* Fix export interaction tab navigation failing test

* Moved a constant functions declaration

Is a best practice to declare constant functions in the beginning of the codebase before calling, in order the javascript interpreter can locate.
This action being reverted and noted as part of flakiness/tech debt.

* Use interaction `subject` rather than export `title`

Amend the current implementation of export interaction list by using `subject` rather than export interaction `title` on the list because is already on the header and repetitive each details .