Hide "Add investment project" button if company has no address
Given that UK based companies should not be creating investment projects, it makes sense to hide the button when the company address is unknown, because we cannot differentiate between domestic and international companies.

This also mitigates against an issue where the add investment project form was raising errors when source companies had no address.